foodiereviews/requirements.txt
djangorestframework
django-authority
django-filter
django-widget-tweaks
